+++ packages/wdio-browser-runner/tests/runner.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from 'vitest'
 
 test('need more tests', () => {
-    expect(true).toBe(false)
+    expect(true).toBe(true)
 })
