+++ packages/wdio-cli/tests/commands/install.test.ts
@@ -76,7 +76,7 @@ You can create one by running 'wdio config'`, undefined)
         expect(console.log).toHaveBeenCalledWith('Installing "wdio-chromedriver-service".')
         expect(console.log).toHaveBeenCalledWith('Package "wdio-chromedriver-service" installed successfully.')
         expect(utils.replaceConfig).toHaveBeenCalled()
-        expect(fs.writeFileSync).toHaveBeenCalled()
+        expect(fs.writeFile).toHaveBeenCalled()
         expect(console.log).toHaveBeenCalledWith('Your wdio.conf.js file has been updated.')
         expect(process.exit).toHaveBeenCalledWith(0)
     })
